Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
requests
python-dotenv
httpx
email-validator
